+++ backend/dataall/modules/dashboards/services/dashboard_service.py
@@ -1,6 +1,6 @@
 from dataall.base.context import get_context
 from dataall.core.activity.db.activity_models import Activity
-from dataall.core.environment.env_permission_checker import has_group_permission
+from dataall.core.permissions.services.group_policy_service import GroupPolicyService
 from dataall.core.environment.services.environment_service import EnvironmentService
 from dataall.core.permissions.services.resource_policy_service import ResourcePolicyService
 from dataall.core.permissions.services.tenant_policy_service import TenantPolicyService
@@ -33,7 +33,7 @@ class DashboardService:
     @staticmethod
     @TenantPolicyService.has_tenant_permission(MANAGE_DASHBOARDS)
     @ResourcePolicyService.has_resource_permission(CREATE_DASHBOARD)
-    @has_group_permission(CREATE_DASHBOARD)
+    @GroupPolicyService.has_group_permission(CREATE_DASHBOARD)
     def import_dashboard(uri: str, admin_group: str, data: dict = None) -> Dashboard:
         context = get_context()
         with context.db_engine.scoped_session() as session:
